+++ src/main/java/insane96mcp/enhancedai/modules/skeleton/SkeletonFleeTarget.java
@@ -1,5 +1,6 @@
 package insane96mcp.enhancedai.modules.skeleton;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.ai.EAAvoidEntityGoal;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.enhancedai.setup.EATags;
@@ -7,11 +8,13 @@ import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
-import insane96mcp.insanelib.base.config.Blacklist;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.data.IdTagMatcher;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.ai.goal.WrappedGoal;
 import net.minecraft.world.entity.monster.AbstractSkeleton;
@@ -19,9 +22,10 @@ import net.minecraft.world.entity.player.Player;
 
 import java.util.List;
 
-@Label(name = "Skeleton Flee", description = "Skeletons try to stay away from the target.")
+@Label(name = "Skeleton Flee", description = "Skeletons try to stay away from the target. Use the enhancedai:skeleton_flee entity type tag to add/remove skeletons that are affected by this feature")
 @LoadFeature(module = Modules.Ids.SKELETON)
 public class SkeletonFleeTarget extends Feature {
+    public static final TagKey<EntityType<?>> SKELETON_FLEE = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "skeleton_flee"));
     @Config(min = 0d, max = 1d)
     @Label(name = "Avoid Player chance", description = "Chance for a Skeleton to spawn with the ability to avoid the player")
     public static Double avoidPlayerChance = 0.5d;
@@ -36,15 +40,10 @@ public class SkeletonFleeTarget extends Feature {
     public static Double fleeDistanceFar = 16d;
     @Config(min = 0d, max = 4d)
     @Label(name = "Flee speed Multiplier Near", description = "Speed multiplier when the skeleton avoids the player and it's within 'Flee Distance Near' blocks from him.")
-    public static Double fleeSpeedNear = 1.5d;
+    public static Double fleeSpeedNear = 1.25d;
     @Config(min = 0d, max = 4d)
     @Label(name = "Flee speed Multiplier Far", description = "Speed multiplier when the skeleton avoids the player and it's farther than 'Flee Distance Far' blocks from him.")
-    public static Double fleeSpeedFar = 1.25d;
-    @Config
-    @Label(name = "Entity Blacklist", description = "Entities that will not be affected by this feature")
-    public static Blacklist entityBlacklist = new Blacklist(List.of(
-            IdTagMatcher.newId("quark:forgotten")
-    ), false);
+    public static Double fleeSpeedFar = 1.1d;
 
     public SkeletonFleeTarget(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
@@ -52,7 +51,8 @@ public class SkeletonFleeTarget extends Feature {
 
     public static void onReassessWeaponGoal(AbstractSkeleton skeleton) {
         if (!isEnabled(SkeletonFleeTarget.class)
-                || entityBlacklist.isEntityBlackOrNotWhitelist(skeleton)) return;
+                || !skeleton.getType().is(SKELETON_FLEE))
+            return;
 
         CompoundTag persistentData = skeleton.getPersistentData();