+++ src/main/java/insane96mcp/enhancedai/modules/villager/VillagerAttacking.java
@@ -1,26 +1,29 @@
 package insane96mcp.enhancedai.modules.villager;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.enhancedai.setup.EATags;
 import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
-import insane96mcp.insanelib.base.config.Blacklist;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-import java.util.List;
-
-@Label(name = "Villager Attacking", description = "Make villagers fight back")
+@Label(name = "Villager Attacking", description = "Make villagers fight back. Use the enhancedai:villagers_can_attack entity type tag to add more villagers.")
 @LoadFeature(module = Modules.Ids.VILLAGER)
 public class VillagerAttacking extends Feature {
+    public static final TagKey<EntityType<?>> VILLAGERS_CAN_ATTACK = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "villagers_can_attack"));
     @Config
     @Label(name = "Villagers Fight back", description = "If true, when attacked, villagers will call other villagers for help and attack back. Attack damage can't be changed via config due to limitation so use mods like Mobs Properties Randomness to change the damage. Base damage is 4")
     public static Boolean villagersFightBack = true;
@@ -32,10 +35,7 @@ public class VillagerAttacking extends Feature {
     public static Integer minReputationFightBack = -100;
     @Config(min = 0d, max = 4d)
     @Label(name = "Movement Speed Multiplier", description = "Movement speed multiplier when attacking")
-    public static Double speedMultiplier = 0.8d;
-    @Config
-    @Label(name = "Entity Blacklist", description = "Entities that shouldn't be affected by this feature")
-    public static Blacklist entityBlacklist = new Blacklist(List.of(), false);
+    public static Double speedMultiplier = 0.4d;
 
     public VillagerAttacking(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
@@ -46,7 +46,7 @@ public class VillagerAttacking extends Feature {
     public void onSpawn(EntityJoinLevelEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntity() instanceof Villager villager)
-                || entityBlacklist.isEntityBlackOrNotWhitelist(villager))
+                || !villager.getType().is(VILLAGERS_CAN_ATTACK))
             return;
 
         CompoundTag persistentData = villager.getPersistentData();