+++ changelog.md
@@ -6,7 +6,9 @@
   * Chance is now saved in the mob (and can be changed)
   * Now drops the item on left or right
 * Misc Illagers removed, any mob can now open doors (by default only Illagers)
+* Added a "bonus movement speed" config option to Movement feature
 * Halved Snow golems snowballs damage
+* Mobs can now mount pigs alongside spiders
 * Many more tags and tags are now always listed in the feature description in the config
 * Launch creepers are now less accurate on hard and more accurate on easy
 * Launch creeper explosion radius override can now be disabled